requirements.txt
yaml
paramiko
python-dotenv